fix: comments on reports
