+++ src/app/(public)/reports/[id]/_components/comment-section.tsx
@@ -1,11 +1,9 @@
 "use client";
 
-import { useEffect, useState, useMemo } from "react";
+import { useState, useMemo, useEffect } from "react";
+import { useRouter } from "next/navigation";
 import { MessageSquare } from "lucide-react";
-import { getReportComments } from "@/actions/comments";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
-import { Button } from "@/components/ui/button";
 import {
   Select,
   SelectContent,
@@ -16,31 +14,33 @@ import {
 import { ReportCommentForm } from "./comment-form";
 import { ReportCommentThread } from "./comment-thread";
 
+type CommentWithMeta = Awaited<
+  ReturnType<typeof import("@/actions/comments").getReportComments>
+>[number];
+
 interface ReportCommentSectionProps {
   reportId: string;
+  initialComments: CommentWithMeta[];
 }
 
 type SortOption = "hot" | "top" | "new" | "controversial";
 
-export function ReportCommentSection({ reportId }: ReportCommentSectionProps) {
-  const [comments, setComments] = useState<any[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
+export function ReportCommentSection({
+  reportId,
+  initialComments,
+}: ReportCommentSectionProps) {
+  const router = useRouter();
+  const [comments, setComments] = useState<CommentWithMeta[]>(initialComments);
   const [sortBy, setSortBy] = useState<SortOption>("hot");
 
+  // Sync comments when initialComments changes (after server refresh)
   useEffect(() => {
-    loadComments();
-  }, [reportId]);
+    setComments(initialComments);
+  }, [initialComments]);
 
-  const loadComments = async () => {
-    setIsLoading(true);
-    try {
-      const data = await getReportComments(reportId);
-      setComments(data);
-    } catch (error) {
-      console.error("Failed to load comments:", error);
-    } finally {
-      setIsLoading(false);
-    }
+  const handleCommentAdded = () => {
+    // Refresh to get updated comments from server
+    router.refresh();
   };
 
   const sortedComments = useMemo(() => {
@@ -118,19 +118,12 @@ export function ReportCommentSection({ reportId }: ReportCommentSectionProps) {
         </div>
       </CardHeader>
       <CardContent className="space-y-6">
-        <ReportCommentForm reportId={reportId} onCommentAdded={loadComments} />
+        <ReportCommentForm
+          reportId={reportId}
+          onCommentAdded={handleCommentAdded}
+        />
 
-        {isLoading ? (
-          <div className="space-y-4">
-            {[1, 2, 3].map((i) => (
-              <div key={i} className="space-y-2">
-                <Skeleton className="h-4 w-3/4" />
-                <Skeleton className="h-4 w-1/2" />
-                <Skeleton className="h-4 w-2/3" />
-              </div>
-            ))}
-          </div>
-        ) : comments.length === 0 ? (
+        {comments.length === 0 ? (
           <p className="text-center text-sm text-muted-foreground py-6">
             No comments yet. Be the first to comment!
           </p>
@@ -141,7 +134,7 @@ export function ReportCommentSection({ reportId }: ReportCommentSectionProps) {
                 key={comment.id}
                 comment={comment}
                 reportId={reportId}
-                onCommentAdded={loadComments}
+                onCommentAdded={handleCommentAdded}
               />
             ))}
           </div>
