+++ src/app/(public)/reports/[id]/_components/report-data-loader.tsx
@@ -1,5 +1,7 @@
 import { getReportById } from "@/actions/reports";
 import { getReportVerifications } from "@/actions/verifications";
+import { getUserVote } from "@/actions/votes";
+import { getSessionWithOnboarding } from "@/lib/session-utils";
 import { notFound } from "next/navigation";
 import { ReportDetailContent } from "./report-detail-content";
 
@@ -16,5 +18,18 @@ export async function ReportDataLoader({ reportId }: ReportDataLoaderProps) {
 
   const verifications = await getReportVerifications(reportId);
 
-  return <ReportDetailContent report={report} verifications={verifications} />;
+  // Get user session and vote
+  const session = await getSessionWithOnboarding();
+  const userVote = session?.user?.id
+    ? await getUserVote(reportId, session.user.id)
+    : null;
+
+  return (
+    <ReportDetailContent
+      report={report}
+      verifications={verifications}
+      userVote={userVote}
+      userId={session?.user?.id ?? null}
+    />
+  );
 }
